feat: show "Waiting for connection..." when sending messages
In order to simplify message sending and failure, we decided to remove
the message sending timeout of 30 seconds and retry forever.

This, however, requires that the user sees a more informative footer
about the state of the message. This commit introduces this new text
simply by checking the network state when a message is in the SENDING
state.