fix: ensure references to await signals are not lost
It seems as though the underlying signals used to complete the future in
the SyncService#await method were being garbage collected before they
reached an isFinished state

I'm still not sure exactly why the underlying objects are being garbage
collected, but introducing a map to hold onto them seems to solve the bug.

My best guess is that in between the time we create the LiveDataSignal
and when the observer gets used after posting the `addObserver` call to
the UI thread, that some reference is lost and gets GCed. This is still
pretty fishy though, since all my tests for Signal#head show that calling
head on a signal prevents it from being GCed, and so the same should
apply here.

In order to reproduce this bug, sign in and wait for the first sync to
finish. Kill the app and re-open it. If it syncs again, the references
to the underlying signal didn't complete on the first attempt and so we
failed to set the `shouldSyncInitial` preference to false.

That is, when you kill and re-open the app, it shouldn't sync again