fix-internal: rename string with modified placeholders
Changing the type of a placeholder (in this case from Int to String) is
never a good idea! The old translations of the string will still be the
old type, but the resource resolution step won't recognise this, and will
pull them in on devices with a different language setting. This means,
that until the translators catch up, running wire in a different language
will crash!

Instead, best to rename the string so that resource resolution fails to
find it.